Support metric keys with multiple metric types
go-metrics allows the same key to be used for multiple metric types
(gauge + histogram for example). Specifically vault exports
vault.database.CreateUser as both a histogram and a counter.

This expands ExtractLabelsFn hook to take the metric type in as a
parameter and will append "counter" or "gauge" to the end of the metric
name to avoid collisions.